+++ openquake/calculators/base.py
