+++ openquake/calculators/calc.py
