+++ openquake/calculators/event_based_risk.py
@@ -23,7 +23,7 @@ import collections
 import numpy
 
 from openquake.baselib.general import AccumDict, groupby
-from openquake.commonlib.calculators import base
+from openquake.calculators import base
 from openquake.commonlib import readinput, parallel, datastore
 from openquake.risklib import riskinput, scientific, workflows
