+++ openquake/calculators/tests/base_test.py
