+++ openquake/commonlib/commands/info.py
@@ -25,7 +25,7 @@ from openquake.baselib.performance import Monitor
 from openquake.baselib.general import humansize, groupby
 from openquake.commonlib import (
     sap, readinput, nrml, source, datastore, reportwriter)
-from openquake.commonlib.calculators import base
+from openquake.calculators import base
 from openquake.risklib import riskinput
 from openquake.hazardlib import gsim
